feat: share core function
