+++ index.js
@@ -2,10 +2,7 @@
 
 const fs = require('fs')
 const path = require('path')
-const lockfile = require('@yarnpkg/lockfile')
-const eol = require('eol')
-const nmtree = require('nmtree')
-const { buildYarnTree, buildNpmTree } = require('./lib/tree')
+const { yarnToNpmCore, npmToYarnCore } = require('./lib')
 
 module.exports = {
   yarnToNpm (packageDir) {
@@ -13,28 +10,15 @@ module.exports = {
       path.join(packageDir, 'yarn.lock'),
       'utf-8'
     )
-    const yarnLockNormalized = eol.lf(yarnLock)
-    const yarnObject = lockfile.parse(yarnLockNormalized).object
-    const nodeModulesTree = nmtree(packageDir)
-    const dependencies = buildNpmTree(nodeModulesTree, yarnObject)
     const packageJson = fs.readFileSync(path.join(packageDir, 'package.json'))
     const { name, version } = JSON.parse(packageJson)
-    return JSON.stringify({
-      name,
-      version,
-      lockfileVersion: 1,
-      requires: true,
-      dependencies
-    })
+    return yarnToNpmCore(yarnLock, name, version, packageDir)
   },
   npmToYarn (packageDir) {
     const packageLockFileString = fs.readFileSync(
       path.join(packageDir, 'package-lock.json'),
       'utf-8'
     )
-    const packageLock = JSON.parse(packageLockFileString)
-    const nodeModulesTree = nmtree(packageDir)
-    const yarnTree = buildYarnTree(nodeModulesTree, packageLock)
-    return lockfile.stringify(yarnTree)
+    return npmToYarnCore(packageLockFileString, packageDir)
   }
 }
